TMC-57 changed data type of externalServiceTimeCf to Double

+++ src/main/java/com/example/timecraft/domain/config/dto/ConfigGetResponse.java
@@ -13,6 +13,6 @@ public class ConfigGetResponse {
   private Integer workingDayEndHour;
   private Boolean isJiraEnabled;
   private Boolean isExternalServiceEnabled;
-  private Integer externalServiceTimeCf;
+  private Double externalServiceTimeCf;
   private Boolean isExternalServiceIncludeDescription;
 }
